Crate bs2x_pac

Peripheral access API for BS2X microcontrollers (generated using svd2rust v0.37.1 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use self::gpio0 as gpio1;

pub use self::gpio0 as gpio2;

pub use self::gpio0 as gpio3;

pub use self::gpio0 as gpio4;

pub use self::gpio0 as ulp_gpio;

pub use self::uart0 as uart1;

pub use self::uart0 as uart2;

pub use self::i2c0 as i2c1;

pub use self::spi0 as spi1;

pub use self::spi0 as spi2;

pub use self::dma as sdma;

Modules

adc_pmu_afe

ADC PMU AFE power/isolation/reset block

aon_afe

AON AFE isolation control used by the BS2X ADC porting layer

dma

DMA controller with 4 channels

gadc

13-bit GADC (general ADC, v153)

generic

Common register and bit access and modify traits

glb_ctl_m

Main core global control - BCPU/MCPU reset status, chip reset, AON CRG

gpio0

GPIO controller for GPIO[7:0]

i2c0

I2C master controller (DesignWare-compatible v151)

interrupt

Interrupt numbers, priority levels, and HART IDs.

keyscan

Key-scan matrix controller (v150)

pdm

PDM microphone interface (v150)

pwm

PWM controller with 8 channels

qdec

Quadrature decoder (v150)

rtc

RTC v150 instance 0 register block

spi0

SPI0 master/slave controller (SSI v151)

tcxo

TCXO 64-bit free-running counter (v150)

timer

Timer module with 3 independent 32-bit timers (v150)

trng

True random number generator (v1)

uart0

UART0 - Universal Asynchronous Receiver/Transmitter

usb

USB 2.0 OTG controller (Synopsys DWC OTG, device-controller base)

wdt

Watchdog timer (v151)

Structs

Peripherals

All the peripherals.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority

Type Aliases

AdcPmuAfe

ADC PMU AFE power/isolation/reset block

AonAfe

AON AFE isolation control used by the BS2X ADC porting layer

Dma

DMA controller with 4 channels

Gadc

13-bit GADC (general ADC, v153)

GlbCtlM

Main core global control - BCPU/MCPU reset status, chip reset, AON CRG

Gpio0

GPIO controller for GPIO[7:0]

Gpio1

GPIO controller for GPIO[7:0]

Gpio2

GPIO controller for GPIO[7:0]

Gpio3

GPIO controller for GPIO[7:0]

Gpio4

GPIO controller for GPIO[7:0]

I2c0

I2C master controller (DesignWare-compatible v151)

I2c1

I2C master controller (DesignWare-compatible v151)

Keyscan

Key-scan matrix controller (v150)

Pdm

PDM microphone interface (v150)

Pwm

PWM controller with 8 channels

Qdec

Quadrature decoder (v150)

Rtc

RTC v150 instance 0 register block

Sdma

DMA controller with 4 channels

Spi0

SPI0 master/slave controller (SSI v151)

Spi1

SPI0 master/slave controller (SSI v151)

Spi2

SPI0 master/slave controller (SSI v151)

Tcxo

TCXO 64-bit free-running counter (v150)

Timer

Timer module with 3 independent 32-bit timers (v150)

Trng

True random number generator (v1)

Uart0

UART0 - Universal Asynchronous Receiver/Transmitter

Uart1

UART0 - Universal Asynchronous Receiver/Transmitter

Uart2

UART0 - Universal Asynchronous Receiver/Transmitter

UlpGpio

GPIO controller for GPIO[7:0]

Usb

USB 2.0 OTG controller (Synopsys DWC OTG, device-controller base)

Wdt

Watchdog timer (v151)