bs2x_pac::generic

Struct Reg

pub struct Reg<REG: RegisterSpec> { /* private fields */ }

This structure provides volatile access to registers.

Implementations

impl<REG: RegisterSpec> Reg<REG>

pub fn as_ptr(&self) -> *mut REG::Ux

Returns the underlying memory address of register.

let reg_ptr = periph.reg.as_ptr();

impl<REG: Readable> Reg<REG>

pub fn read(&self) -> R<REG>

Reads the contents of a Readable register.

You can read the raw contents of a register by using bits:

let bits = periph.reg.read().bits();

or get the content of a particular field of a register:

let reader = periph.reg.read();
let bits = reader.field1().bits();
let flag = reader.field2().bit_is_set();

impl<REG: Resettable + Writable> Reg<REG>

pub fn reset(&self)

Writes the reset value to Writable register.

Resets the register to its initial state.

pub fn write<F>(&self, f: F) -> REG::Ux
where F: FnOnce(&mut W<REG>) -> &mut W<REG>,

Writes bits to a Writable register.

You can write raw bits into a register:

periph.reg.write(|w| unsafe { w.bits(rawbits) });

or write only the fields you need:

periph.reg.write(|w| w
    .field1().bits(newfield1bits)
    .field2().set_bit()
    .field3().variant(VARIANT)
);

or an alternative way of saying the same:

periph.reg.write(|w| {
    w.field1().bits(newfield1bits);
    w.field2().set_bit();
    w.field3().variant(VARIANT)
});

In the latter case, other fields will be set to their reset value.

pub fn from_write<F, T>(&self, f: F) -> T
where F: FnOnce(&mut W<REG>) -> T,

Writes bits to a Writable register and produce a value.

You can write raw bits into a register:

periph.reg.write_and(|w| unsafe { w.bits(rawbits); });

or write only the fields you need:

periph.reg.write_and(|w| {
    w.field1().bits(newfield1bits)
        .field2().set_bit()
        .field3().variant(VARIANT);
});

or an alternative way of saying the same:

periph.reg.write_and(|w| {
    w.field1().bits(newfield1bits);
    w.field2().set_bit();
    w.field3().variant(VARIANT);
});

In the latter case, other fields will be set to their reset value.

Values can be returned from the closure:

let state = periph.reg.write_and(|w| State::set(w.field1()));

impl<REG: Writable> Reg<REG>

pub unsafe fn write_with_zero<F>(&self, f: F) -> REG::Ux
where F: FnOnce(&mut W<REG>) -> &mut W<REG>,

Writes 0 to a Writable register.

Similar to write, but unused bits will contain 0.

Safety

Unsafe to use with registers which don’t allow to write 0.

pub unsafe fn from_write_with_zero<F, T>(&self, f: F) -> T
where F: FnOnce(&mut W<REG>) -> T,

Writes 0 to a Writable register and produces a value.

Similar to write, but unused bits will contain 0.

Safety

Unsafe to use with registers which don’t allow to write 0.

impl<REG: Readable + Writable> Reg<REG>

pub fn modify<F>(&self, f: F) -> REG::Ux
where for<'w> F: FnOnce(&R<REG>, &'w mut W<REG>) -> &'w mut W<REG>,

Modifies the contents of the register by reading and then writing it.

E.g. to do a read-modify-write sequence to change parts of a register:

periph.reg.modify(|r, w| unsafe { w.bits(
   r.bits() | 3
) });

or

periph.reg.modify(|_, w| w
    .field1().bits(newfield1bits)
    .field2().set_bit()
    .field3().variant(VARIANT)
);

or an alternative way of saying the same:

periph.reg.modify(|_, w| {
    w.field1().bits(newfield1bits);
    w.field2().set_bit();
    w.field3().variant(VARIANT)
});

Other fields will have the value they had before the call to modify.

pub fn from_modify<F, T>(&self, f: F) -> T
where for<'w> F: FnOnce(&R<REG>, &'w mut W<REG>) -> T,

Modifies the contents of the register by reading and then writing it and produces a value.

E.g. to do a read-modify-write sequence to change parts of a register:

let bits = periph.reg.modify(|r, w| {
    let new_bits = r.bits() | 3;
    unsafe {
        w.bits(new_bits);
    }

    new_bits
});

or

periph.reg.modify(|_, w| {
    w.field1().bits(newfield1bits)
        .field2().set_bit()
        .field3().variant(VARIANT);
});

or an alternative way of saying the same:

periph.reg.modify(|_, w| {
    w.field1().bits(newfield1bits);
    w.field2().set_bit();
    w.field3().variant(VARIANT);
});

Other fields will have the value they had before the call to modify.