bs2x_pac::interrupt

Function cause

pub fn cause() -> Trap

Retrieves the cause of a trap in the current hart (machine mode).

If the raw cause is not a valid interrupt or exception for the target, it panics.